ime fixes - add composing region to presenter
